separate ui module for home POC1 Ref:#47
