+++ .idea/misc.xml
@@ -43,7 +43,17 @@
         <entry key="../../../../layout/compose-model-1646549400146.xml" value="0.5" />
         <entry key="../../../../layout/compose-model-1646549400147.xml" value="0.46111111111111114" />
         <entry key="../../../../layout/compose-model-1646549572214.xml" value="0.375" />
-        <entry key="../../../../layout/compose-model-1646556208340.xml" value="0.25" />
+        <entry key="../../../../layout/compose-model-1647256502643.xml" value="0.46111111111111114" />
+        <entry key="../../../../layout/compose-model-1649241193416.xml" value="0.46111111111111114" />
+        <entry key="../../../../layout/compose-model-1649241553618.xml" value="0.2504222972972973" />
+        <entry key="../../../../layout/compose-model-1649243592015.xml" value="0.2504222972972973" />
+        <entry key="../../../../layout/compose-model-1649943826816.xml" value="0.234375" />
+        <entry key="../../../../layout/compose-model-1649943900416.xml" value="0.39537037037037037" />
+        <entry key="../../../../layout/compose-model-1650018986876.xml" value="0.1" />
+        <entry key="../../../../layout/compose-model-1650019147052.xml" value="0.24197635135135134" />
+        <entry key="../../../../layout/compose-model-1650019355945.xml" value="0.12249163879598662" />
+        <entry key="../../../../layout/compose-model-1650019373381.xml" value="0.12249163879598662" />
+        <entry key="../../../../layout/compose-model-1650019511264.xml" value="0.24788851351351351" />
         <entry key="app/src/main/res/layout/layout_main.xml" value="0.11820652173913043" />
       </map>
     </option>
