+++ app/src/main/res/navigation/nav_graph.xml
@@ -25,7 +25,7 @@
         android:label="PlayerFragment" >
         <argument
             android:name="dataItem"
-            app:argType="com.imdmp.ui_home.VideoListItem"
+            app:argType="com.imdmp.uihome.VideoListItem"
             app:nullable="true"
             android:defaultValue="@null"/>
     </fragment>
