+++ ui-home/build.gradle
@@ -1,6 +1,9 @@
 plugins {
     id 'com.android.application'
     id 'kotlin-android'
+    id 'kotlin-kapt'
+    id 'kotlin-parcelize'
+
 }
 
 android {
@@ -29,6 +32,14 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
+
+    buildFeatures {
+        compose true
+    }
+
+    composeOptions {
+        kotlinCompilerExtensionVersion compose_version
+    }
 }
 
 dependencies {
@@ -43,7 +54,12 @@ dependencies {
     implementation "androidx.compose.ui:ui:$compose_version"
     implementation "androidx.compose.material:material:$compose_version"
     implementation "androidx.compose.ui:ui-tooling:$compose_version"
+    implementation 'androidx.activity:activity-compose:1.3.1'
 
+    kapt 'com.github.bumptech.glide:compiler:4.12.0'
+    implementation "com.github.skydoves:landscapist-glide:1.4.8"
+    implementation "androidx.constraintlayout:constraintlayout-compose:1.0.0"
 
+    implementation("br.com.devsrsouza.compose.icons.android:font-awesome:1.0.0")
 
 }
