+++ ui-home/src/main/java/com/imdmp/uihome/MainActivity.kt
@@ -1,0 +1,18 @@
+package com.imdmp.uihome
+
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.material.MaterialTheme
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        setContent {
+            MaterialTheme {
+                HomeScreen()
+            }
+        }
+    }
+}
