+++ ui-home/src/main/java/com/imdmp/uihome/VideoItemActions.kt
@@ -1,11 +1,11 @@
-package com.imdmp.ui_home
+package com.imdmp.uihome
 
 interface VideoItemActions {
-    fun videoItemSelected(videoListItem: com.imdmp.ui_home.VideoListItem)
+    fun videoItemSelected(videoListItem: com.imdmp.uihome.VideoListItem)
 
     companion object {
         fun default(): VideoItemActions = object : VideoItemActions {
-            override fun videoItemSelected(videoListItem: com.imdmp.ui_home.VideoListItem) {
+            override fun videoItemSelected(videoListItem: com.imdmp.uihome.VideoListItem) {
                 TODO("Not yet implemented")
             }
