+++ ui-home/src/main/java/com/imdmp/uihome/topappbar/HomeTopAppBar.kt
@@ -1,4 +1,4 @@
-package com.imdmp.ui_home.topappbar
+package com.imdmp.uihome.topappbar
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
@@ -10,10 +10,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.imdmp.youtubecompose.R
 import compose.icons.FontAwesomeIcons
 import compose.icons.fontawesomeicons.Brands
 import compose.icons.fontawesomeicons.Solid
@@ -40,7 +38,7 @@ fun HomeTopAppBar(
             )
         },
         title = {
-            Text(text = stringResource(id = R.string.app_name), color = Color.Black)
+            Text(text = "Sample title", color = Color.Black)
         },
 
         actions = {
