+++ ui-home/src/main/java/com/imdmp/uihome/topappbar/ToolbarActions.kt
@@ -1,4 +1,4 @@
-package com.imdmp.ui_home.topappbar
+package com.imdmp.uihome.topappbar
 
 interface ToolbarActions {
     fun searchClicked()
