+++ ui-home/src/test/java/com/imdmp/uihome/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
-package com.imdmp.ui_home
+package com.imdmp.uihome
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
